🐛 fix(tts): Remove async selectors
